fix IPFS deployer subdirectory duplication issue, refactor IPFS deployer to be unit-testable
