+++ packages/cli/src/lib/defaults/deploy-modules/ipfs/index.ts
@@ -4,70 +4,11 @@ import { Uri } from "@polywrap/core-js";
 import { PolywrapClient } from "@polywrap/client-js";
 import * as Sys from "@polywrap/sys-config-bundle-js";
 import { PolywrapClientConfigBuilder } from "@polywrap/client-config-builder-js";
-import fs from "fs";
+import { AddResult, ArgsAddDir, DirectoryEntry } from "./types";
+import { readDirContents } from "./utils";
 
 const isValidUri = (uri: Uri) =>
   uri.authority === "fs" || uri.authority === "file";
-
-interface FileEntry {
-  name: string;
-  data: Uint8Array;
-}
-
-interface DirectoryEntry {
-  name: string;
-  directories?: DirectoryEntry[];
-  files?: FileEntry[];
-}
-
-interface AddOptions {
-  pin?: boolean;
-  onlyHash?: boolean;
-  wrapWithDirectory?: boolean;
-}
-
-interface AddResult {
-  name: string;
-  hash: string;
-  size: string;
-}
-
-interface ArgsAddDir {
-  data: DirectoryEntry;
-  ipfsProvider: string;
-  timeout?: number;
-  addOptions?: AddOptions;
-}
-
-const readDirContents = async (
-  path: string,
-  dirName: string
-): Promise<DirectoryEntry> => {
-  const dirents: fs.Dirent[] = await fs.promises.readdir(path, {
-    withFileTypes: true,
-  });
-  const data: DirectoryEntry = { name: dirName };
-
-  for (const dirent of dirents) {
-    if (dirent.isDirectory()) {
-      const subDir = await readDirContents(
-        `${path}/${dirent.name}`,
-        `${dirName}/${dirent.name}`
-      );
-      data.directories = data.directories ?? [];
-      data.directories?.push(subDir);
-    } else {
-      const fileData = await fs.promises.readFile(`${path}/${dirent.name}`);
-      data.files = data.files ?? [];
-      data.files?.push({
-        name: dirent.name,
-        data: fileData,
-      });
-    }
-  }
-
-  return data;
-};
 
 class IPFSDeployer implements DeployModule {
   async execute(uri: Uri, config?: { gatewayUri: string }): Promise<Uri> {
